First Commit
I've submitted Sync to the GBJam forums and it's done. I wasn't sure if
I wanted to put the source code up but I don't really see a reason not
to anymore.

I should note that a lot of my code is not commented at all so please
don't message me about how it's not commented.