Changed Player Input Choices + More comments
I stepped away from this code for a few days and now I have updated it.
The first thing I did was look into some gameplay feedback I recieved.
It seems that having to constantly to press the X button is annoying. So
the player no longer needs to press the X button to light up the tiles.
They merely have to move over an unlit tile for it to switch. Also the
player can hold down an Arrow Key to move instead of having the press
them constantly.
I am also thinking about putting in a series of "Instruction" screens so
when the player plays for the first time they get a run down of the
controls.